refactor: Create account directory in entities
